Componentized code; fixed #27
