+++ src/lib/components/program/sprint/Content.svelte
@@ -1,0 +1,24 @@
+<script>
+    import Heading from '$lib/components/Heading.svelte'
+
+    export let title, sprintNumber, content
+</script>
+
+
+<section class="blue-on-green pilled col-span-2">
+    <Heading title="Sprint {sprintNumber}: {title}" />
+
+    <div>
+        {#if content}
+        {@html content.html}
+        {/if}
+    </div>
+</section>
+
+
+<style>
+    section.col-span-2 {
+        max-width:42rem;
+        padding-bottom: 1rem;
+    }
+</style>
