+++ src/lib/components/program/sprint/Topics.svelte
@@ -1,0 +1,16 @@
+<!-- <aside>
+    <h3>Topics</h3>
+    <ul>
+    <li class="green-on-blue rounded">
+        
+        {#if startdate}
+            <IconCal date={longDate(startdate)} />
+        {/if}
+    </li>
+    {#each tags as tag}
+        <li class="green-on-blue rounded">
+        <IconTag {tag} />
+        </li>
+    {/each}
+    </ul>
+</aside> -->
