+++ src/routes/+layout.svelte
@@ -20,6 +20,9 @@
 <Footer />
 
 <style>
+	:global(body) {
+      display: block !important;
+    }
 	:global(body:before){
 		left:auto;
 		top:auto;
@@ -30,22 +33,22 @@
 	:global(body:after){
 		left:1rem
 	}
-
+	main {
+      display: block !important;
+    }
 	h1 {
 		margin-left: 1rem;
 		color:transparent
 	}
-
 	h1 span {
 		position: absolute;
-		font-size: 7rem;
-		top: .5rem;
+		font-size: 6rem;
+		top: 1.65rem;
 		color: rgb(166, 117, 245);
 		z-index: -1;
 		left: 1.5rem;
 		transition:.25s
 	}
-
 	:global(footer nav) {
 		padding:1rem;
 	}
@@ -56,7 +59,7 @@
 	@media (min-width: 30rem) {
 		h1 span {
 			font-size: 10rem;
-			top: -2.5rem;
+			top: -1.75rem;
 			left:1.25rem
 		}
     }
